+++ CHANGELOG.md
@@ -2,11 +2,21 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.11.6] - 2025-04-04
+
+### 🐛 Bug Fixes
+
+- Backend request authority
+
 ## [0.11.5] - 2025-04-03
 
 ### 🐛 Bug Fixes
 
 - Mapping store name with db url
+
+### ⚙️ Miscellaneous Tasks
+
+- Release
 
 ## [0.11.4] - 2025-04-02